requirements.txt
Werkzeug==2.0.3+computecanada
xmltodict==0.12.0+computecanada
yq==2.14.0
zipp==3.8.0+computecanada
absl-py==1.0.0+computecanada
argcomplete==1.12.3+computecanada
cachetools==4.2.4+computecanada
certifi==2021.10.8+computecanada
charset-normalizer==2.0.12+computecanada
cloudpickle==1.6.0+computecanada
cycler==0.11.0+computecanada
fonttools==4.32.0+computecanada
future==0.18.2+computecanada
google-auth==2.3.3+computecanada
google-auth-oauthlib==0.4.6+computecanada
grpcio==1.44.0+computecanada
gym==0.18.0+computecanada
idna==3.3+computecanada
importlib-metadata==4.11.3+computecanada
kiwisolver==1.3.2+computecanada
markdown==3.3.6+computecanada
matplotlib==3.5.1+computecanada
numpy==1.22.2+computecanada
oauthlib==3.1.1+computecanada
packaging==21.3+computecanada
Pillow==7.2.0+computecanada
protobuf==3.19.4+computecanada
pyasn1==0.4.8+computecanada
pyasn1-modules==0.2.8+computecanada
pyglet==1.5.0
pyparsing==3.0.8+computecanada
python-dateutil==2.8.2+computecanada
pyyaml==6.0+computecanada
requests==2.27.1+computecanada
requests-oauthlib==1.3.0+computecanada
rsa==4.8+computecanada
scipy==1.8.0+computecanada
six==1.16.0+computecanada
tensorboard==2.8.0+computecanada
tensorboard-data-server==0.6.1+computecanada
tensorboard-plugin-wit==1.8.0+computecanada
toml==0.10.2+computecanada
torch==1.10.0+computecanada
typing-extensions==4.1.1+computecanada
urllib3==1.26.9+computecanada

